+++ redturtle/imagedevent/__init__.py
@@ -1,17 +1,16 @@
 """Main product initializer
 """
 
+import logging
+
 from zope.i18nmessageid import MessageFactory
 from redturtle.imagedevent import config
 
 from Products.Archetypes import atapi
 from Products.CMFCore import utils
 
-# Define a message factory for when this product is internationalised.
-# This will be imported with the special name "_" in most modules. Strings
-# like _(u"message") will then be extracted by i18n tools for translation.
-
 imagedeventMessageFactory = MessageFactory('redturtle.imagedevent')
+logger = logging.getLogger('redturtle.imagedevent')
 
 def initialize(context):
     """Initializer called when used as a Zope 2 product.